Error handling based on the http status code
fixed #21
